// src/utils/types.tsx
export interface AllCharacter {
    info: {
        count: number
        pages: number
        next: string
        prev?: any
    },

    results: characterProps[]
}

export interface characterProps {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: {
        name: string;
        url: string;
    };
    location: {
        name: string;
        url: string;
    };
    image: string;
    episode: string[];
    url: string;
    created: string;
}